+++ lib/nextjs-themes/src/client/theme-switcher/theme-switcher.tsx
@@ -1,10 +1,7 @@
-import * as React from "react";
 import { useEffect } from "react";
-import { resolveTheme, parseState, encodeState } from "../../utils";
-import { ColorSchemeType, DEFAULT_ID, ThemeStoreType, initialState } from "../../constants";
-import useRGS, { SetStateAction } from "r18gs";
+import { resolveTheme, MEDIA, useStore } from "../../utils";
+import { ColorSchemeType, DARK, DEFAULT_ID, LIGHT, ResolvedColorSchemeType } from "../../constants";
 
-/** todo - set persistance and cookies */
 export interface ThemeSwitcherProps {
   forcedTheme?: string;
   forcedColorScheme?: ColorSchemeType;
@@ -14,51 +11,18 @@ export interface ThemeSwitcherProps {
   styles?: Record<string, string>;
 }
 
-function useMediaQuery(setThemeState: SetStateAction<ThemeStoreType>) {
-  React.useEffect(() => {
-    // set event listener for media
-    const media = matchMedia("(prefers-color-scheme: dark)");
-    const updateSystemColorScheme = () => {
-      setThemeState(state => ({ ...state, systemColorScheme: media.matches ? "dark" : "light" }));
-    };
-    updateSystemColorScheme();
-    media.addEventListener("change", updateSystemColorScheme);
-    return () => {
-      media.removeEventListener("change", updateSystemColorScheme);
-    };
-  }, [setThemeState]);
-}
-
-let tInit = 0;
-function useLoadSyncedState(setThemeState: SetStateAction<ThemeStoreType>, targetSelector?: string) {
-  React.useEffect(() => {
-    tInit = Date.now();
-    const key = targetSelector ?? DEFAULT_ID;
-    const storedState = parseState(localStorage.getItem(key));
-    // @ts-ignore
-    delete storedState.systemColorScheme;
-    setThemeState(state => ({ ...state, ...storedState }));
-    const storageListener = (e: StorageEvent) => {
-      if (e.key === key) setThemeState(state => ({ ...state, ...parseState(e.newValue) }));
-    };
-    window.addEventListener("storage", storageListener);
-    return () => {
-      window.removeEventListener("storage", storageListener);
-    };
-  }, [targetSelector]);
-}
-
 export interface DataProps {
   className: string;
   "data-th"?: string;
   "data-theme"?: string;
-  "data-color-scheme"?: "dark" | "light";
-  "data-csp"?: ColorSchemeType /** color-scheme-preference */;
+  "data-color-scheme"?: ResolvedColorSchemeType;
+  /** color-scheme-preference */
+  "data-csp"?: ColorSchemeType;
 }
 
 export interface UpdateProps {
   resolvedTheme: string;
-  resolvedColorScheme: "dark" | "light";
+  resolvedColorScheme: ResolvedColorSchemeType;
   resolvedColorSchemePref: ColorSchemeType;
   th: string;
 }
@@ -84,22 +48,18 @@ const updateDOM = (
   return shouldCreateCookie;
 };
 
-const disableAnimation = (themeTransition = "none") => {
+/** disable transition while switching theme */
+const disableTransition = (themeTransition = "none") => {
   const css = document.createElement("style");
   /** split by ';' to prevent CSS injection */
-  const transition = `transition: ${themeTransition.split(";")[0]} !important;`;
-  css.appendChild(
-    document.createTextNode(`*{-webkit-${transition}-moz-${transition}-o-${transition}-ms-${transition}${transition}}`),
-  );
+  css.textContent = `transition: ${themeTransition.split(";")[0]} !important;`;
   document.head.appendChild(css);
 
   return () => {
     // Force restyle
-    (() => window.getComputedStyle(document.body))();
+    getComputedStyle(document.body);
     // Wait for next tick before removing
-    setTimeout(() => {
-      document.head.removeChild(css);
-    }, 1);
+    setTimeout(() => document.head.removeChild(css), 1);
   };
 };
 
@@ -107,31 +67,37 @@ const disableAnimation = (themeTransition = "none") => {
  * You can use this hook in place of `<ThemeSwitcher />` component.
  * Please note that you need to add "use client" on top of the component in which you are using this hook.
  */
-export function useThemeSwitcher(props: ThemeSwitcherProps) {
-  const [themeState, setThemeState] = useRGS<ThemeStoreType>(props.targetSelector ?? DEFAULT_ID, initialState);
-
-  useMediaQuery(setThemeState);
-  useLoadSyncedState(setThemeState, props.targetSelector);
+export const useThemeSwitcher = (props: ThemeSwitcherProps) => {
+  const [themeState, setThemeState] = useStore(props.targetSelector);
+  // not using ?? as we don't want key to be an empty string ever
+  const key = props.targetSelector || DEFAULT_ID;
+  /** set listeners for system preference and syncing store */
   useEffect(() => {
-    const restoreTransitions = disableAnimation(props.themeTransition);
+    const media = matchMedia(MEDIA);
+    media.addEventListener("change", () =>
+      setThemeState(state => ({ ...state, systemColorScheme: media.matches ? DARK : LIGHT })),
+    );
+    addEventListener("storage", e => {
+      if (e.key === key) setThemeState(state => ({ ...state, ...JSON.parse(e.newValue || "{}") }));
+    });
+  }, []);
 
+  useEffect(() => {
+    const restoreTransitions = disableTransition(props.themeTransition);
     const resolvedData = resolveTheme(themeState, props);
     const shouldCreateCookie = updateDOM(resolvedData, props);
-    if (tInit < Date.now() - 300) {
-      const stateStr = encodeState(themeState);
-      const key = props.targetSelector || DEFAULT_ID;
-      localStorage.setItem(key, stateStr);
-      if (shouldCreateCookie) document.cookie = `${key}=${stateStr}; max-age=31536000; SameSite=Strict;`;
-    }
+    const stateStr = JSON.stringify(themeState);
+    localStorage.setItem(key, stateStr);
+    if (shouldCreateCookie) document.cookie = `${key}=${stateStr};max-age=31536000;SameSite=Strict;`;
     restoreTransitions();
   }, [props, themeState]);
-}
+};
 
 /**
  * Use this component in your layout - `app/layout.tsx` or your custom layout or in `_app.tsx` file.
  * @component
  */
-export function ThemeSwitcher(props: ThemeSwitcherProps) {
+export const ThemeSwitcher = (props: ThemeSwitcherProps) => {
   useThemeSwitcher(props);
   return null;
-}
+};