+++ lib/nextjs-themes/src/server/nextjs/server-side-wrapper/server-side-wrapper.test.tsx
@@ -2,7 +2,6 @@ import { cleanup, render, screen } from "@testing-library/react";
 import { afterEach, describe, test, beforeEach } from "vitest";
 import { NextJsSSGThemeSwitcher, ServerSideWrapper } from ".";
 import { DEFAULT_ID } from "../../../constants";
-import { encodeState } from "../../../utils";
 
 describe("server-side-target", () => {
   afterEach(cleanup);
@@ -10,12 +9,12 @@ describe("server-side-target", () => {
   beforeEach(() => {
     globalThis.cookies = {
       [DEFAULT_ID]: {
-        value: encodeState({
+        value: JSON.stringify({
           theme: "yellow",
           darkTheme: "dark-blue",
           lightTheme: "light-yellow",
           colorSchemePref: "dark",
-          systemColorScheme: "dark"
+          systemColorScheme: "dark",
         }),
       },
     };