+++ lib/nextjs-themes/src/utils.ts
@@ -1,51 +1,44 @@
+import useRGS from "r18gs";
 import { ThemeSwitcherProps, UpdateProps } from "./client";
-import { ColorSchemeType, ThemeStoreType, initialState } from "./constants";
+import { DARK, DEFAULT_ID, LIGHT, ResolvedColorSchemeType, SYSTEM, ThemeStoreType, initialState } from "./constants";
 
-export function resolveTheme(state?: ThemeStoreType, props?: ThemeSwitcherProps): UpdateProps {
-  const resolvedForcedTheme = props?.forcedTheme === undefined ? state?.forcedTheme : props.forcedTheme;
-  const resolvedForcedColorScheme =
-    props?.forcedColorScheme === undefined ? state?.forcedColorScheme : props.forcedColorScheme;
-  const resolvedColorSchemePref =
-    (resolvedForcedColorScheme === undefined ? state?.colorSchemePref : resolvedForcedColorScheme) || "";
+/** resolve props and state to a final attributes that should be applied to the DOM */
+export const resolveTheme = (state?: ThemeStoreType, props?: ThemeSwitcherProps): UpdateProps => {
+  const resolvedForcedTheme = props?.forcedTheme ?? state?.forcedTheme;
+  const resolvedForcedColorScheme = props?.forcedColorScheme ?? state?.forcedColorScheme;
+  const resolvedColorSchemePref = (resolvedForcedColorScheme ?? state?.colorSchemePref) || "";
 
-  const isSystemDark = state?.systemColorScheme === "dark";
+  const isSystemDark = state?.systemColorScheme === DARK;
 
-  let resolvedColorScheme: "dark" | "light" = isSystemDark ? "dark" : "light";
-  let resolvedTheme = resolvedForcedTheme === undefined ? state?.theme || "" : resolvedForcedTheme;
+  /** these will be modified in the switch statement */
+  let resolvedColorScheme: ResolvedColorSchemeType = isSystemDark ? DARK : LIGHT;
+  let resolvedTheme = (resolvedForcedTheme ?? state?.theme) || "";
 
   if (resolvedForcedTheme === undefined)
     switch (resolvedColorSchemePref) {
-      case "system":
+      case SYSTEM:
         resolvedTheme = (isSystemDark ? state?.darkTheme : state?.lightTheme) || "";
         break;
-      case "dark":
-        [resolvedTheme, resolvedColorScheme] = [state?.darkTheme || "", "dark"];
+      case DARK:
+        [resolvedTheme, resolvedColorScheme] = [state?.darkTheme || "", DARK];
         break;
-      case "light":
-        [resolvedTheme, resolvedColorScheme] = [state?.lightTheme || "", "light"];
+      case LIGHT:
+        [resolvedTheme, resolvedColorScheme] = [state?.lightTheme || "", LIGHT];
         break;
     }
 
   const th = resolvedForcedTheme === undefined ? state?.theme || "" : resolvedForcedTheme;
   return { resolvedTheme, resolvedColorScheme, resolvedColorSchemePref, th };
-}
+};
 
-export function getResolvedTheme() {
-  return document.documentElement.getAttribute("data-theme");
-}
+const isServer = typeof localStorage === "undefined";
+export const MEDIA = "(prefers-color-scheme: dark)";
 
-export function getResolvedColorScheme() {
-  return document.documentElement.getAttribute("data-color-scheme");
-}
-
-export function encodeState(themeState: ThemeStoreType) {
-  const { colorSchemePref, systemColorScheme, darkTheme, lightTheme, theme } = themeState;
-  return [colorSchemePref, systemColorScheme, darkTheme, lightTheme, theme].join(",");
-}
-
-export function parseState(str?: string | null): ThemeStoreType {
-  if(!str) return initialState;
-  type StrSplitType = [ColorSchemeType, "dark" | "light", string, string, string];
-  const [colorSchemePref, systemColorScheme, darkTheme, lightTheme, theme] = str.split(",") as StrSplitType;
-  return { colorSchemePref, systemColorScheme, darkTheme, lightTheme, theme };
-}
+/** internal store API */
+export const useStore = (targetId?: string) => {
+  const key = targetId ?? DEFAULT_ID;
+  return useRGS<ThemeStoreType>(key, () => {
+    const str = isServer ? null : localStorage.getItem(key);
+    return str ? { ...JSON.parse(str), systemColorScheme: matchMedia(MEDIA).matches ? DARK : LIGHT } : initialState;
+  });
+};